+++ src/vue/App.vue
@@ -81,6 +81,7 @@ export default {
   methods: {
     ...mapActions({
       loadKvEntries: vuexTypes.LOAD_KV_ENTRIES,
+      loadDefaultQuoteAsset: vuexTypes.LOAD_DEFAULT_QUOTE_ASSET,
     }),
     async initApp () {
       await Sdk.init(config.HORIZON_SERVER)
@@ -90,6 +91,7 @@ export default {
         Sdk.sdk.useWallet(this[vuexTypes.wallet])
         Api.useWallet(this[vuexTypes.wallet])
         await this.loadKvEntries()
+        await this.loadDefaultQuoteAsset()
       }
     },
     detectIE () {